backend/pyproject.toml
[project]
name = "coinhq-backend"
version = "0.1.0"
description = "CoinHQ — Multi-profile crypto portfolio tracker backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "pydantic==2.7.1",
    "pydantic-settings==2.3.0",
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "redis[asyncio]==5.0.4",
    "cryptography==42.0.8",
    "httpx==0.27.0",
    "slowapi==0.1.9",
    "python-multipart==0.0.9",
    "python-binance==1.0.19",
    "pybit==5.6.2",
    "python-okx>=0.3",
    "python-jose[cryptography]>=3.3.0",
    "authlib>=1.3.0",
    "structlog>=24.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "httpx>=0.27.0",
    "aiosqlite>=0.20",
    "ruff>=0.4.0",
    "pip-audit>=2.7",
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "pytest-mock", "aiosqlite", "httpx"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
